+++ scalyr_agent/agent_main.py
@@ -106,7 +106,7 @@ from scalyr_agent.scalyr_client import create_client, verify_server_certificate
 from scalyr_agent.copying_manager import CopyingManager
 from scalyr_agent.configuration import Configuration
 from scalyr_agent.util import RunState, ScriptEscalator
-from scalyr_agent.util import warn_on_python26
+from scalyr_agent.util import warn_on_old_or_unsupported_python_version
 from scalyr_agent.agent_status import AgentStatus
 from scalyr_agent.agent_status import ConfigStatus
 from scalyr_agent.agent_status import OverallStats
@@ -302,7 +302,7 @@ class ScalyrAgent(object):
         @return:  The exit status code to exit with, such as 0 for success.
         @rtype: int
         """
-        warn_on_python26()
+        warn_on_old_or_unsupported_python_version()
 
         quiet = command_options.quiet
         verbose = command_options.verbose